requirements.txt
flask
flask-limiter
flask-cors
flask-wtf
python-dotenv
mysql-connector-python
docker>=6.0.0
pyjwt
bcrypt